dev_book/guess_my_age/requirements.txt
flask==2.3.3
opencv-python-headless==4.8.0.76
numpy==1.24.3
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5
werkzeug==2.3.7